+++ amv-hm-access-sdk/src/main/java/org/amv/access/sdk/hm/certificate/LocalStorage.java
@@ -27,4 +27,6 @@ public interface LocalStorage {
     Observable<AccessCertificatePair> findAccessCertificates();
 
     Observable<Boolean> removeAccessCertificateById(String accessCertificateId);
+
+    Observable<Boolean> reset();
 }