+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Remove command timeouts when disconnecting
+- Reset local storage on errors during initialization
 
 ## [0.1.0] - 2018-03-06
 ### Added